refactor(upload): restructure file upload and navigation components
