+++ service/app/src/entities/game/ui/components/index.ts
@@ -1,0 +1,4 @@
+export { CreateGameNavigation } from "./createGameNavigation"
+export { FileUploadArea } from "./fileUploadArea"
+export { QuestionInputForm } from "./questionInputForm"
+export { QuestionList } from "./questionList"
