+++ service/app/src/entities/game/ui/components/questionList.tsx
