+++ service/app/src/entities/game/ui/popups/index.ts
@@ -1,0 +1,4 @@
+export { FileSizeErrorPopup } from "./fileSizeErrorPopup"
+export { FileTypeErrorPopup } from "./fileTypeErrorPopup"
+export { LibraryRegisterPopup } from "./libraryRegisterPopup" 
+export { SaveConfirmPopup } from "./saveConfirmPopup"
